ThechessApplication: allow Notifiable to be created or removed during notifying
Set of waiting Notifiable's and pointer to active Object
were added to ThechessApplication.
If Notifiable was created or deleted during notifying,
it would be added or erased from this set.
Pointer to Object is used to find out whether created/deleted
Notifiable belongs to notifing Object.
Outside of ThechessApplication::thechess_notify()
pointer to Object points to 0 -- it is indicator of
thechess_notify execution().

model/Object: add operator==