Makefile: fix style target
Astyle considers *_FOREACH a function. So
 * astyle removes a space after *_FOREACH
 * astyle adds a space before * and & inside *_FOREACH argument list
This behavior was fixed with two sed commands

Note: THECHESS_MOVES[_FROM][_TO] were also affected by this bug.
They were also fixed

Astyle was asked to be quiet because its list of changed file may be
incorrect as a result of sed scripts