Competition::create_games_classical(): avoid same pairs if possible
Let it be two candidates for black in game in which white player is
already selected. Both candidates play same number of games with white player
where candidate is black and white is white.

Then select that candidate, who play fewer games with white player
(regardless who is white and black in those games).

see #118